♻️ Extract preview classes to own files
